+++ src/Akka.Persistence.SqlServer.Tests/DbUtils.cs
@@ -1,29 +1,27 @@
-﻿//-----------------------------------------------------------------------
+﻿// -----------------------------------------------------------------------
 // <copyright file="DbUtils.cs" company="Akka.NET Project">
-//     Copyright (C) 2009-2016 Typesafe Inc. <http://www.typesafe.com>
-//     Copyright (C) 2013-2016 Akka.NET project <https://github.com/akkadotnet/akka.net>
+//      Copyright (C) 2013 - 2019 .NET Foundation <https://github.com/akkadotnet/akka.net>
 // </copyright>
-//-----------------------------------------------------------------------
+// -----------------------------------------------------------------------
 
-using System.Configuration;
 using System.Data.SqlClient;
 
 namespace Akka.Persistence.SqlServer.Tests
 {
     public static class DbUtils
     {
+        public static string ConnectionString { get; private set; }
 
-        public static void Initialize()
+        public static void Initialize(string connectionString)
         {
-            var connectionString = ConfigurationManager.ConnectionStrings["TestDb"].ConnectionString;
             var connectionBuilder = new SqlConnectionStringBuilder(connectionString);
 
             //connect to postgres database to create a new database
             var databaseName = connectionBuilder.InitialCatalog;
             connectionBuilder.InitialCatalog = "master";
-            connectionString = connectionBuilder.ToString();
+            ConnectionString = connectionBuilder.ToString();
 
-            using (var conn = new SqlConnection(connectionString))
+            using (var conn = new SqlConnection(ConnectionString))
             {
                 conn.Open();
 
@@ -42,15 +40,18 @@ namespace Akka.Persistence.SqlServer.Tests
                 }
 
                 DropTables(conn, databaseName);
+
+                // set this back to the journal/snapshot database
+                connectionBuilder.InitialCatalog = databaseName;
+                ConnectionString = connectionBuilder.ToString();
             }
         }
 
         public static void Clean()
         {
-            var connectionString = ConfigurationManager.ConnectionStrings["TestDb"].ConnectionString;
-            var connectionBuilder = new SqlConnectionStringBuilder(connectionString);
+            var connectionBuilder = new SqlConnectionStringBuilder(ConnectionString);
             var databaseName = connectionBuilder.InitialCatalog;
-            using (var conn = new SqlConnection(connectionString))
+            using (var conn = new SqlConnection(ConnectionString))
             {
                 conn.Open();
                 DropTables(conn, databaseName);
@@ -61,12 +62,11 @@ namespace Akka.Persistence.SqlServer.Tests
         {
             using (var cmd = new SqlCommand())
             {
-                cmd.CommandText = string.Format(@"
-                    USE {0};
+                cmd.CommandText = $@"
+                    USE {databaseName};
                     IF EXISTS (SELECT * FROM INFORMATION_SCHEMA.TABLES WHERE TABLE_SCHEMA = 'dbo' AND TABLE_NAME = 'EventJournal') BEGIN DROP TABLE dbo.EventJournal END;
                     IF EXISTS (SELECT * FROM INFORMATION_SCHEMA.TABLES WHERE TABLE_SCHEMA = 'dbo' AND TABLE_NAME = 'Metadata') BEGIN DROP TABLE dbo.Metadata END;
-                    IF EXISTS (SELECT * FROM INFORMATION_SCHEMA.TABLES WHERE TABLE_SCHEMA = 'dbo' AND TABLE_NAME = 'SnapshotStore') BEGIN DROP TABLE dbo.SnapshotStore END;",
-                databaseName);
+                    IF EXISTS (SELECT * FROM INFORMATION_SCHEMA.TABLES WHERE TABLE_SCHEMA = 'dbo' AND TABLE_NAME = 'SnapshotStore') BEGIN DROP TABLE dbo.SnapshotStore END;";
                 cmd.Connection = conn;
                 cmd.ExecuteNonQuery();
             }