+++ src/Akka.Persistence.SqlServer/Extension.cs
@@ -1,9 +1,8 @@
-﻿//-----------------------------------------------------------------------
+﻿// -----------------------------------------------------------------------
 // <copyright file="Extension.cs" company="Akka.NET Project">
-//     Copyright (C) 2009-2016 Lightbend Inc. <http://www.lightbend.com>
-//     Copyright (C) 2013-2016 Akka.NET project <https://github.com/akkadotnet/akka.net>
+//      Copyright (C) 2013 - 2019 .NET Foundation <https://github.com/akkadotnet/akka.net>
 // </copyright>
-//-----------------------------------------------------------------------
+// -----------------------------------------------------------------------
 
 using Akka.Actor;
 using Akka.Configuration;
@@ -14,7 +13,7 @@ namespace Akka.Persistence.SqlServer
     public class SqlServerJournalSettings : JournalSettings
     {
         public const string ConfigPath = "akka.persistence.journal.sql-server";
-        
+
         public SqlServerJournalSettings(Config config) : base(config)
         {
         }
@@ -23,38 +22,24 @@ namespace Akka.Persistence.SqlServer
     public class SqlServerSnapshotSettings : SnapshotStoreSettings
     {
         public const string ConfigPath = "akka.persistence.snapshot-store.sql-server";
-        
+
         public SqlServerSnapshotSettings(Config config) : base(config)
         {
         }
     }
 
     /// <summary>
-    /// An actor system extension initializing support for SQL Server persistence layer.
+    ///     An actor system extension initializing support for SQL Server persistence layer.
     /// </summary>
     public class SqlServerPersistence : IExtension
     {
         /// <summary>
-        /// Returns a default configuration for akka persistence SQLite-based journals and snapshot stores.
-        /// </summary>
-        /// <returns></returns>
-        public static Config DefaultConfiguration()
-        {
-            return ConfigurationFactory.FromResource<SqlServerPersistence>("Akka.Persistence.SqlServer.sql-server.conf");
-        }
-
-        public static SqlServerPersistence Get(ActorSystem system)
-        {
-            return system.WithExtension<SqlServerPersistence, SqlServerPersistenceProvider>();
-        }
-
-        /// <summary>
-        /// Journal-related settings loaded from HOCON configuration.
+        ///     Journal-related settings loaded from HOCON configuration.
         /// </summary>
         public readonly Config DefaultJournalConfig;
 
         /// <summary>
-        /// Snapshot store related settings loaded from HOCON configuration.
+        ///     Snapshot store related settings loaded from HOCON configuration.
         /// </summary>
         public readonly Config DefaultSnapshotConfig;
 
@@ -66,10 +51,25 @@ namespace Akka.Persistence.SqlServer
             DefaultJournalConfig = defaultConfig.GetConfig(SqlServerJournalSettings.ConfigPath);
             DefaultSnapshotConfig = defaultConfig.GetConfig(SqlServerSnapshotSettings.ConfigPath);
         }
+
+        /// <summary>
+        ///     Returns a default configuration for akka persistence SQLite-based journals and snapshot stores.
+        /// </summary>
+        /// <returns></returns>
+        public static Config DefaultConfiguration()
+        {
+            return ConfigurationFactory.FromResource<SqlServerPersistence>(
+                "Akka.Persistence.SqlServer.sql-server.conf");
+        }
+
+        public static SqlServerPersistence Get(ActorSystem system)
+        {
+            return system.WithExtension<SqlServerPersistence, SqlServerPersistenceProvider>();
+        }
     }
 
     /// <summary>
-    /// Singleton class used to setup SQL Server backend for akka persistence plugin.
+    ///     Singleton class used to setup SQL Server backend for akka persistence plugin.
     /// </summary>
     public class SqlServerPersistenceProvider : ExtensionIdProvider<SqlServerPersistence>
     {