+++ RELEASE_NOTES.md
@@ -1,3 +1,36 @@
+#### 1.3.14 July 30 2019 ####
+* Upgrades to Akka.Persistence v1.3.14 - resolves major issues with Akka.Cluster.Sharding serialization.
+* [BatchingSqlJournal now preserves Sender in PersistCallback](https://github.com/akkadotnet/akka.net/pull/3779)
+
+#### 1.3.13 April 30 2019 ####
+* Upgrades to Akka.Persistence v1.3.13
+* Resolves https://github.com/akkadotnet/Akka.Persistence.SqlServer/issues/104 - major issue with BatchingSqlJournal.
+
+#### 1.3.7 June 04 2018 ####
+Upgrades to Akka.Persistence 1.3.7, which includes some major changes for SQL-based SnapShot stores, which you can read more about here: https://github.com/akkadotnet/akka.net/issues/3422
+
+This should significantly improve the performance of SQL-based journals for loading large snapshots.
+
+Also, removes the dependency on the Akka.TestKit for this package.
+
+
+#### 1.3.2 October 29 2017 ####
+
+Updated to Akka.Persistence v1.3.2. Fixed bug in SnapshotStore query.
+
+#### 1.3.1 September 11 2017 ####
+
+Support for Akka.NET 1.3, .NET Standard 1.6, and the first stable RTM release of Akka.Persistence.
+
+**Migration from 1.1.1.7-beta Up**
+
+The event journal and snapshot store schema has changed with this release.  In order to keep existing stores compatible with this release, you **must** add a column to both stores for `SerializerId` like so:
+
+```sql
+ALTER TABLE {your_journal_table_name} ADD COLUMN SerializerId INTEGER NULL
+ALTER TABLE {your_snapshot_table_name} ADD COLUMN SerializerId INTEGER NULL
+```
+
 #### 1.1.1 August 01 2016 ####
 Support for Akka.NET 1.1 and Akka.Persistence.Query